renamed error handling module to fit standards
